coyote_v2/requirements.txt
# COYOTE V2.5 Dependencies

# Core
anthropic>=0.18.0
click>=8.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# SMS Alerts
twilio>=8.0.0

# Testing
pytest>=7.0.0

# Optional: FastAPI for webhook server
fastapi>=0.100.0
uvicorn>=0.20.0
